gh-2347 Switched to sequential safe info loading
